# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangle API Documentation",
        "description": "API documentation for Tangle application. This API provides endpoints for managing collaborations, user profiles, reviews, and authentication.",
        "termsOfService": "https://tangle-api.example.com/terms",
        "contact": {
            "name": "Tangle API Support",
            "url": "https://tangle-api.example.com",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get user's reports",
                "operationId": "42398321972bc3074eef76718e0fd9dc",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's reports"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Create a new report",
                "operationId": "318de4ef8123b483320dc9dbb45b6a43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reportable_type",
                                    "reportable_id",
                                    "reason"
                                ],
                                "properties": {
                                    "reportable_type": {
                                        "type": "string",
                                        "enum": [
                                            "collaboration_request",
                                            "user",
                                            "message",
                                            "review",
                                            "application"
                                        ]
                                    },
                                    "reportable_id": {
                                        "type": "integer"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "enum": [
                                            "spam",
                                            "scam",
                                            "offensive",
                                            "fake_opportunity",
                                            "inappropriate",
                                            "harassment",
                                            "other"
                                        ]
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get report details",
                "operationId": "a4c119b5a8302fa439721ad28809a1fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report details"
                    },
                    "404": {
                        "description": "Report not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/reports": {
            "get": {
                "tags": [
                    "Admin Reports"
                ],
                "summary": "Get all reports (admin only)",
                "operationId": "07327546f93ac360a523eecc9b535f96",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all reports"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/reports/{id}/review": {
            "post": {
                "tags": [
                    "Admin Reports"
                ],
                "summary": "Review a report (admin only)",
                "operationId": "7474c042c99aa070648ffc5da9c7c114",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "enum": [
                                            "approve",
                                            "dismiss"
                                        ]
                                    },
                                    "admin_notes": {
                                        "type": "string"
                                    },
                                    "admin_action": {
                                        "type": "string",
                                        "enum": [
                                            "none",
                                            "warn",
                                            "suspend",
                                            "delete",
                                            "hide"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report reviewed successfully"
                    },
                    "404": {
                        "description": "Report not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/reports/{id}/mark-review": {
            "post": {
                "tags": [
                    "Admin Reports"
                ],
                "summary": "Mark report as under review (admin only)",
                "operationId": "1d1f0269722af77cc427c8bb7c2f48b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report marked as under review"
                    },
                    "404": {
                        "description": "Report not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/reports/stats": {
            "get": {
                "tags": [
                    "Admin Reports"
                ],
                "summary": "Get report statistics (admin only)",
                "operationId": "bfda012d19ea17e70f84d9c4f3923ae1",
                "responses": {
                    "200": {
                        "description": "Report statistics"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
...
}